removed scale because i felt like it, also ui has separate resolution now

+++ 2D game template/2D game/Game.cpp
@@ -125,7 +125,7 @@ bool Game::Update()
 	}
 
 	frameMs = time.Read();
-	timeToDelay = (1.0f / (float)frameCap * 1000) - frameMs;
+	timeToDelay = (int)(1.0f / (float)frameCap * 1000.0f) - frameMs;
 	
 	if(timeToDelay > 0)
 	{
